gomock: refactor directory name
Rename directory to match default package output of gomock.
